requirements.txt
pillow
opencv-python
tqdm
torch
torchvision
numpy
onnxruntime

flask